# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Small KMS Admin API
  version: 0.1.0
servers:
  - url: http://localhost:9000
    description: Local server
paths:
  /v2/{namespaceId}/certificate-templates/{templateId}/certificates/latest:
    parameters:
      - $ref: "#/components/parameters/NamespaceIdParameter"
      - $ref: "#/components/parameters/TemplateIdParameter"
      - $ref: "#/components/parameters/IncludeCertificateParameter"
    get:
      tags:
        - admin
      summary: Get certificate
      operationId: GetLatestCertificateByTemplateV2
      responses:
        200:
          $ref: "#/components/responses/CertificateResponse"
  /v2/{namespaceId}/certificate-templates/{templateId}/enroll:
    parameters:
      - $ref: "#/components/parameters/NamespaceIdParameter"
      - $ref: "#/components/parameters/TemplateIdParameter"
    post:
      tags:
        - enroll
      summary: Put certificate template
      operationId: BeginEnrollCertificateV2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateEnrollmentRequest"
        required: true
      responses:
        201:
          description: Certificate template response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateEnrollmentReceipt"
        400:
          $ref: "#/components/responses/ErrorResponse"
  /v2/{namespaceId}/link-service-principal:
    parameters:
      - $ref: "#/components/parameters/NamespaceIdParameter"
    get:
      tags:
        - admin
      summary: Link device service principal
      operationId: GetDeviceServicePrincipalLinkV2
      responses:
        200:
          description: Link device service principal response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicePrincipalLinkedDevice"
    post:
      tags:
        - admin
      summary: Link device service principal
      operationId: CreateDeviceServicePrincipalLinkV2
      responses:
        201:
          description: Link device service principal response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicePrincipalLinkedDevice"
        400:
          $ref: "#/components/responses/ErrorResponse"
        409:
          $ref: "#/components/responses/ErrorResponse"
  /v2/{namespaceId}/certificates/{certId}:
    parameters:
      - $ref: "#/components/parameters/NamespaceIdParameter"
      - $ref: "#/components/parameters/CertIdParameter"
      - $ref: "#/components/parameters/IncludeCertificateParameter"
    get:
      tags:
        - admin
      summary: Get certificate
      operationId: GetCertificateV2
      responses:
        200:
          $ref: "#/components/responses/CertificateResponse"
  /v2/{namespaceId}/certificates/{certId}/pending:
    parameters:
      - $ref: "#/components/parameters/NamespaceIdParameter"
      - $ref: "#/components/parameters/CertIdParameter"
      - $ref: "#/components/parameters/IncludeCertificateParameter"
    post:
      tags:
        - enroll
      summary: complete certificate enrollment
      operationId: CompleteCertificateEnrollmentV2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateEnrollmentReplyFinalize"
        required: true
      responses:
        200:
          $ref: "#/components/responses/CertificateResponse"
        400:
          $ref: "#/components/responses/ErrorResponse"
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    NamespaceIdParameter:
      in: path
      name: namespaceId
      schema:
        type: string
        format: uuid
      required: true
...
    TemplateIdParameter:
      in: path
      name: templateId
      schema:
        type: string
        format: uuid
      required: true
    CertIdParameter:
      in: path
      name: certId
      schema:
        type: string
        format: uuid
      required: true
    IncludeCertificateParameter:
      in: query
      name: "includeCertificate"
      schema:
        $ref: "#/components/schemas/IncludeCertificate"
  schemas:
    IncludeCertificate:
      type: string
      enum:
        - jwk
        - pem
      x-enum-varnames:
        - IncludeJWK
        - IncludePEM
...
    RefType:
      type: string
      enum:
        - namespace
        - certificate-template
        - certificate
        - certificate-enrollment-receipt
      x-enum-varnames:
        - RefTypeNamespace
        - RefTypeCertificateTemplate
        - RefTypeCertificate
        - RefTypeCertificateEnrollReceipt
    Ref:
      type: object
      properties:
        namespaceId:
          type: string
          format: uuid
          x-go-name: NamespaceID
        type:
          $ref: "#/components/schemas/RefType"
        id:
          type: string
          format: uuid
          x-go-name: ID
      required:
        - namespaceId
        - type
        - id
    RefWithMetadata:
      allOf:
        - $ref: "#/components/schemas/Ref"
        - type: object
          properties:
            updatedBy:
              description: Unique ID of the user who last updated the object
              type: string
            updated:
              description: Time when the object was last updated
              type: string
              format: date-time
            deleted:
              description: Time when the object was deleted
              type: string
              format: date-time
            displayName:
              description: Display name of the object
              type: string
            isDefault:
              type: boolean
            isActive:
              type: boolean
          x-go-type-skip-optional-pointer: true
          required:
            - updated
            - updatedBy
            - displayName
...
    CertificateEnrollmentTargetType:
      type: string
      enum:
        - device-linked-service-principal
      x-enum-varnames:
        - CertEnrollTargetTypeDeviceLinkedServicePrincipal
    CertificateEnrollmentRequestDeviceLinkedServicePrincipal:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/CertificateEnrollmentTargetType"
        deviceNamespaceId:
          description: Object ID of the device
          type: string
          format: uuid
          x-go-name: DeviceNamespaceID
        linkId:
          description: Unique ID of the device link
          type: string
          format: uuid
          x-go-name: DeviceLinkID
        appId:
          description: Client ID of the application
          type: string
          format: uuid
          x-go-name: AppID
        servicePrincipalId:
          description: Object ID of the service principal
          type: string
          format: uuid
          x-go-name: ServicePrincipalID
        commonName:
          description: Common Name to appear in the certificate
          type: string
      required:
        - type
        - deviceNamespaceId
        - linkId
        - appId
        - servicePrincipalId
        - fqdn
    CertificateEnrollmentRequest:
      oneOf:
        - $ref: "#/components/schemas/CertificateEnrollmentRequestDeviceLinkedServicePrincipal"
      discriminator:
        propertyName: type
        mapping:
          device-linked-service-principal: "#/components/schemas/CertificateEnrollmentRequestDeviceLinkedServicePrincipal"
    CertificateEnrollmentReceipt:
      type: object
      properties:
        ref:
          $ref: "#/components/schemas/RefWithMetadata"
        expires:
          description: Time when the enrollment expires
          type: string
          format: date-time
        jwtClaims:
          description: payload section of the certificate claims, in JWT format, base64url encoded
          type: string
        templateNamespaceId:
          description: Unique ID of the namespace of the certificate template
          type: string
          format: uuid
          x-go-name: TemplateNamespaceID
        templateId:
          description: Consistent derived ID (UUID v5) of the certificate template
          type: string
          format: uuid
          x-go-name: TemplateID
        requesterId:
          description: Unique ID of the user who requested the certificate
          type: string
          format: uuid
          x-go-name: RequesterID
        keyProperties:
          $ref: "#/components/schemas/JwkProperties"
      required:
        - ref
        - expires
        - jwtClaims
        - templateNamespaceId
        - templateId
        - requesterId
        - keyProperties
    CertificateEnrollmentReplyFinalize:
      type: object
      properties:
        jwtHeader:
          description: header section of the certificate claims, in JWT format, base64url encoded
          type: string
        jwtSignature:
          description:
            signature section of the jwt, serves as proof of confirmation finalize the enrollment and being issued a certificate,
            with header and signature, signed with the key pair with the public key in the same request
          type: string
        publicKey:
          $ref: "#/components/schemas/JwkProperties"
      required:
        - jwtHeader
        - jwtSignature
        - publicKey
...
    ServicePrincipalLinkedDevice:
      type: object
      properties:
        ref:
          $ref: "#/components/schemas/RefWithMetadata"
        deviceOid:
          description: Object ID of the device
          type: string
          format: uuid
          x-go-name: DeviceOID
        deviceId:
          type: string
          format: uuid
          x-go-name: DeviceID
        applicationOid:
          description: Object ID of the application
          type: string
          format: uuid
          x-go-name: ApplicationOID
        applicationClientId:
          type: string
          format: uuid
          x-go-name: ApplicationClientID
        servicePrincipalId:
          type: string
          format: uuid
          x-go-name: ServicePrincipalID
        status:
          type: string
      required:
        - ref
        - deviceOid
        - deviceId
        - applicationClientId
        - applicationOid
        - servicePrincipalId
        - status
...
    CertificateUsage:
      type: string
      enum:
        - root-ca
        - intermediate-ca
        - server-and-client
        - server-only
        - client-only
        - aad-client-credential
      x-enum-varnames:
        - Usage_RootCA
        - Usage_IntCA
        - Usage_ServerAndClient
        - Usage_ServerOnly
        - Usage_ClientOnly
        - Usage_AADClientCredential
...
    CertificateSubjectAlternativeNames:
      type: object
      properties:
        emails:
          x-go-name: EmailAddresses
          type: array
          items:
            type: string
          x-go-type-skip-optional-pointer: true
        uris:
          x-go-name: URIs
          type: array
          items:
            type: string
          x-go-type-skip-optional-pointer: true
...
    CertificateInfo:
      type: object
      properties:
        ref:
          $ref: "#/components/schemas/RefWithMetadata"
        commonName:
          description: Common name
          type: string
        subject:
          type: string
        subjectAlternativeNames:
          $ref: "#/components/schemas/CertificateSubjectAlternativeNames"
        notBefore:
          description: Expiration date of the certificate
          type: string
          format: date-time
        notAfter:
          description: Expiration date of the certificate
          type: string
          format: date-time
        usage:
          $ref: "#/components/schemas/CertificateUsage"
        jwk:
          $ref: "#/components/schemas/JwkProperties"
        pem:
          type: string
        template:
          $ref: "#/components/schemas/Ref"
        issuerCertificate:
          $ref: "#/components/schemas/Ref"
      required:
        - ref
        - commonName
        - subject
        - notBefore
        - notAfter
        - usage
        - content
        - template
        - issuerCertificate
...
    JwkAlg:
      type: string
      enum:
        - RS256
        - RS384
        - RS512
        - ES256
        - ES384
      x-enum-varnames:
        - AlgRS256
        - AlgRS384
        - AlgRS512
        - AlgES256
        - AlgES384
    JwtKty:
      type: string
      enum:
        - RSA
        - EC
      x-enum-varnames:
        - KeyTypeRSA
        - KeyTypeEC
    JwtCrv:
      type: string
      enum:
        - P-256
        - P-384
      x-enum-varnames:
        - CurveNameP256
        - CurveNameP384
    JwkKeyOperation:
      type: string
      enum:
        - sign
        - verify
        - encrypt
        - decrypt
        - wrapKey
        - unwrapKey
      x-enum-varnames:
        - KeyOpSign
        - KeyOpVerify
        - KeyOpEncrypt
        - KeyOpDecrypt
        - KeyOpWrapKey
        - KeyOpUnwrapKey
      x-go-name: KeyOp
    JwkProperties:
      type: object
      description: Property bag of JSON Web Key (RFC 7517) with additional fields, all bytes are base64url encoded
      properties:
        alg:
          $ref: "#/components/schemas/JwkAlg"
        kty:
          $ref: "#/components/schemas/JwtKty"
        key_size:
          description: RSA key size
          type: integer
        key_ops:
          $ref: "#/components/schemas/JwkKeyOperation"
        n:
          description: RSA modulus
          type: string
        e:
          description: RSA exponent
          type: string
        crv:
          $ref: "#/components/schemas/JwtCrv"
        x:
          description: EC x coordinate
          type: string
        y:
          description: EC y coordinate
          type: string
        kid:
          description: Key ID
          type: string
          x-go-name: KeyID
        x5u:
          x-go-name: CertificateURL
          description: X.509 certificate URL
          type: string
        x5c:
          description: X.509 certificate chain
          type: array
          items:
            type: string
          x-go-name: CertificateChain
          x-go-type-skip-optional-pointer: true
        x5t:
          description: X.509 certificate SHA-1 thumbprint
          type: string
          x-go-name: CertificateThumbprint
        x5t#S256:
          description: X.509 certificate SHA-256 thumbprint
          type: string
          x-go-name: CertificateThumbprintSHA256
      required:
        - kty
  responses:
...
    CertificateResponse:
      description: CertificateResponse
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CertificateInfo"
...
    ErrorResponse:
      description: error response
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true